Fix to #33183 - Complex property causes "The multi-part identifier "xxx" could not be bound" after GroupBy + Select(x => x.First())
Problem was that for group by we clone shaper expression and in case we cloned a table, we assign new aliases to all the columns associated with that table. However, we were not updating TableAlias values on StructuralTypeProjectionExpression. Later, when we try to generate columns for complex properties (in GenerateComplexPropertyShaperExpression) we were using incorrect alias (original one, rather than updated).
Fix is to manually update TableMap when cloning StructuralTypeProjectionExpression

Fixes #33183